+++ src/ui/activities/auth/bookmarks.rs
@@ -32,7 +32,7 @@ use crate::system::environment;
 
 // Ext
 use std::path::PathBuf;
-use tuirealm::components::{input::InputPropsBuilder, radio::RadioPropsBuilder};
+use tui_realm_stdlib::{input::InputPropsBuilder, radio::RadioPropsBuilder};
 use tuirealm::{Payload, PropsBuilder, Value};
 
 impl AuthActivity {