+++ CHANGELOG.md
@@ -22,7 +22,17 @@
 
 ## 0.6.1
 
+Released on ??
 
+- Enhancements:
+  - Now that tui-rs supports title alignment, UI has been improved
+- Dependencies:
+  - Updated `bitflags` to `1.3.2`
+  - Updated `bytesize` to `1.1.0`
+  - Updated `crossterm` to `0.20`
+  - Updated `open` to `2.0.1`
+  - Added `tui-realm-stdlib 0.6.0`
+  - Updated `tui-realm` to `0.6.0`
 
 ## 0.6.0